JQuery Events Calendar Plugin - Metro.js Scheduler
Recently I stumbled across Sergey Pimenov’s  Metro UI framework that
makes the bootstrap framework look a lot like Windows 8 with the tiles
and Microsoft color pallet and a few really nice looking UI elements
that I believe mirror Microsoft components. I really liked how the
schedule streamer presented event data in a linear fashion on tracks,
but I spent a great deal of the last five years working with Sencha’s
Extjs and Touch and I was looking for something that I could drop in
and manipulate the actual contents via passing a JSON object, that way
I could spare myself the tedious job of hand coding a ton of HTML and
also pull data from the database a bit easier. I ended up gutting the
original component and using a variety of other elements to get the
desired functionality and true to the open source license I’m
publishing the finished result below. The reason that the actual
framework is not a dependency is due to the fact that the CSS was
altered so that it did not require adding a class to the body tag; this
allows you to just use the widget without interfering with your
existing styles.